+++ server/services/index.js
@@ -5,6 +5,8 @@ const shop = require('./shop');
 const shopify = require('./shopify');
 const utils = require('./utils');
 const webhooks = require('./webhooks');
+const lifecycles = require('./lifecycles');
+const billing = require('./billing');
 
 module.exports = {
   session,
@@ -12,4 +14,6 @@ module.exports = {
   shopify,
   utils,
   webhooks,
+  lifecycles,
+  ...billing,
 };