+++ server/services/webhooks.js
@@ -1,15 +1,17 @@
 'use strict';
 
+const billingWebhooks = require('./billing/webhooks');
+
 module.exports = ({ strapi }) => ({
   SHOP_UPDATE: async (topic, shop, body) => {
     const shopService = strapi.service('plugin::shopify.shop');
     await shopService.update(body.id, body);
   },
-  // "SHOP_REDACT": (topic, shop, body) => {
 
-  // },
   APP_UNINSTALLED: async (topic, shop, body) => {
     const shopService = strapi.service('plugin::shopify.shop');
     await shopService.uninstall(shop);
   },
+
+  ...billingWebhooks({ strapi }),
 });