+++ CHANGELOG.md
@@ -2,27 +2,56 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [0.0.4](https://github.com/strapify/strapi-plugin-shopify/compare/v0.0.3...v0.0.4) (2022-07-06)
-
-### [0.0.3](https://github.com/strapify/strapi-plugin-shopify/compare/v0.0.2...v0.0.3) (2022-06-27)
+## [0.1.0](https://github.com/shop3/strapi-plugin-shopify/compare/v0.0.6...v0.1.0) (2022-11-03)
 
 
-### Bug Fixes
+### ⚠ BREAKING CHANGES
 
-* **release:** added missing files in npm package ([3e19737](https://github.com/strapify/strapi-plugin-shopify/commit/3e197370cd515c9724d54ab54cea505818064e09))
+* added shop and session to ctx.state.shopify to reduce database calls
 
-### [0.0.2](https://github.com/strapify/strapi-plugin-shopify/compare/v0.0.1...v0.0.2) (2022-06-21)
+### Features
+
+* added shop and session to ctx.state.shopify to reduce database calls ([e0a5e6a](https://github.com/shop3/strapi-plugin-shopify/commit/e0a5e6a43f46febaf69f5ea31573c05574bdd586))
+* cancel subscription after shop is uninstalled ([8d9e450](https://github.com/shop3/strapi-plugin-shopify/commit/8d9e4509ddfa039fa0ebe7e4a9238aa650b909f0))
+
+### [0.0.6](https://github.com/shop3/strapi-plugin-shopify/compare/v0.0.5...v0.0.6) (2022-11-02)
 
 
 ### Features
 
-* added current user endpoint ([af76cf6](https://github.com/strapify/strapi-plugin-shopify/commit/af76cf61d39f6fc3939f433a7d1eabdbc58e21fa))
-* added logout endpoint ([af5b752](https://github.com/strapify/strapi-plugin-shopify/commit/af5b75272901c164a6a0f236f55250ccd8e5e1b0))
-* **settings:** created admin settings page and api endpoint ([8b26e27](https://github.com/strapify/strapi-plugin-shopify/commit/8b26e27ad131fc8fe427cf7c4e940de161e91fd8))
+* added usage record price on db ([0b7f582](https://github.com/shop3/strapi-plugin-shopify/commit/0b7f5824481e2ef6937ae62de4ef9da43b912b43))
+
+### [0.0.5](https://github.com/shop3/strapi-plugin-shopify/compare/v0.0.4...v0.0.5) (2022-11-01)
+
+
+### Features
+
+* added shopify lifecycles ([3198df3](https://github.com/shop3/strapi-plugin-shopify/commit/3198df37ca2c218dba8bc780eedf18ad9936537c))
+* added valid-header-hmac policy ([1ab02a4](https://github.com/shop3/strapi-plugin-shopify/commit/1ab02a43b4b80e8056b552134686877b06283376))
+* initial version of billing feature ([d077154](https://github.com/shop3/strapi-plugin-shopify/commit/d07715408a495d2c5263d65d6ef97109caf8fbcf))
+* return shopify-api from shopify service ([abe20e0](https://github.com/shop3/strapi-plugin-shopify/commit/abe20e0088c0ddd3f2eab1bb04adb516a3b64108))
+
+### [0.0.4](https://github.com/shop3/strapi-plugin-shopify/compare/v0.0.3...v0.0.4) (2022-07-06)
+
+### [0.0.3](https://github.com/shop3/strapi-plugin-shopify/compare/v0.0.2...v0.0.3) (2022-06-27)
+
+
+### Bug Fixes
+
+* **release:** added missing files in npm package ([3e19737](https://github.com/shop3/strapi-plugin-shopify/commit/3e197370cd515c9724d54ab54cea505818064e09))
+
+### [0.0.2](https://github.com/shop3/strapi-plugin-shopify/compare/v0.0.1...v0.0.2) (2022-06-21)
+
+
+### Features
+
+* added current user endpoint ([af76cf6](https://github.com/shop3/strapi-plugin-shopify/commit/af76cf61d39f6fc3939f433a7d1eabdbc58e21fa))
+* added logout endpoint ([af5b752](https://github.com/shop3/strapi-plugin-shopify/commit/af5b75272901c164a6a0f236f55250ccd8e5e1b0))
+* **settings:** created admin settings page and api endpoint ([8b26e27](https://github.com/shop3/strapi-plugin-shopify/commit/8b26e27ad131fc8fe427cf7c4e940de161e91fd8))
 
 ### 0.0.1 (2022-05-27)
 
 
 ### Features
 
-* initial commit of strapi-plugin-shopify ([2e4715d](https://github.com/strapify/strapi-plugin-shopify/commit/2e4715d2670fdf1c394adfe2bf1928c90f7a04e6))
+* initial commit of strapi-plugin-shopify ([2e4715d](https://github.com/shop3/strapi-plugin-shopify/commit/2e4715d2670fdf1c394adfe2bf1928c90f7a04e6))